uda_medical_imbalance_project/requirements.txt
# 核心机器学习框架
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0

# 深度学习框架
torch>=1.9.0
torchvision>=0.10.0

# TabPFN和自动机器学习
tabpfn>=0.1.4

# 域适应库
adapt-python>=0.4.0

# 不平衡数据处理
imbalanced-learn>=0.8.0

# 可视化
matplotlib>=3.4.0
seaborn>=0.11.0
plotly>=5.3.0

# 数据处理
openpyxl>=3.0.7
xlrd>=2.0.1

# 最优传输
POT>=0.8.0

# 贝叶斯优化
scikit-optimize>=0.9.0
hyperopt>=0.2.5

# 距离度量
scipy>=1.7.0

# 实验跟踪
mlflow>=1.20.0
wandb>=0.12.0

# 配置管理
pyyaml>=5.4.0
hydra-core>=1.1.0

# 测试框架
pytest>=6.2.0
pytest-cov>=2.12.0

# 代码质量
black>=21.6.0
flake8>=3.9.0
isort>=5.9.0

# 进度条
tqdm>=4.62.0

# 日志
loguru>=0.5.3

# Jupyter支持
jupyter>=1.0.0
ipywidgets>=7.6.0 